+++ docs/src/examples/QasGridGenerator/Slots.vue
@@ -3,11 +3,34 @@
   <qas-single-view v-model:fields="fields" v-model:result="result" :custom-id="customId" :entity="entity">
     <template #default>
       <div>
-        field-[nome-da-chave]:
         <qas-grid-generator :fields="fields" :result="result">
-          <template #field-name="{ field }">
+          <!-- <template #field-name="{ field }">
             <div>
               name: <pre>{{ field }}</pre>
+            </div>
+          </template> -->
+
+          <template #header-field-name="{ field }">
+            <div>
+              <pre>{{ field }}</pre>
+            </div>
+          </template>
+
+          <template #header>
+            <div>
+              header 2
+            </div>
+          </template>
+
+          <template #content-field-name>
+            <div>
+              content
+            </div>
+          </template>
+
+          <template #content>
+            <div>
+              content 2
             </div>
           </template>
         </qas-grid-generator>