+++ ui/src/components/form-generator/QasFormGenerator.vue
@@ -9,19 +9,21 @@
         </slot>
 
         <div>
-          <div :class="classes">
-            <div v-for="(field, key) in fieldsetItem.fields.visible" :key="key" :class="getFieldClass({ index: key, fields: normalizedFields })">
+          <slot :fields="fieldsetItem.fields?.visible" :name="`legend-section-${fieldsetItemKey}`">
+            <div :class="classes">
+              <div v-for="(field, key) in fieldsetItem.fields.visible" :key="key" :class="getFieldClass({ index: key, fields: normalizedFields })">
+                <slot :field="field" :name="`field-${field.name}`">
+                  <qas-field :disable="isFieldDisabled(field)" v-bind="props.fieldsProps[field.name]" :error="props.errors[key]" :field="field" :model-value="props.modelValue[field.name]" @update:model-value="updateModelValue({ key: field.name, value: $event })" />
+                </slot>
+              </div>
+            </div>
+
+            <div v-for="(field, key) in fieldsetItem.fields.hidden" :key="key">
               <slot :field="field" :name="`field-${field.name}`">
-                <qas-field :disable="isFieldDisabled(field)" v-bind="props.fieldsProps[field.name]" :error="props.errors[key]" :field="field" :model-value="props.modelValue[field.name]" @update:model-value="updateModelValue({ key: field.name, value: $event })" />
+                <qas-field :disable="isFieldDisabled(field)" v-bind="props.fieldsProps[field.name]" :field="field" :model-value="props.modelValue[field.name]" @update:model-value="updateModelValue({ key: field.name, value: $event })" />
               </slot>
             </div>
-          </div>
-
-          <div v-for="(field, key) in fieldsetItem.fields.hidden" :key="key">
-            <slot :field="field" :name="`field-${field.name}`">
-              <qas-field :disable="isFieldDisabled(field)" v-bind="props.fieldsProps[field.name]" :field="field" :model-value="props.modelValue[field.name]" @update:model-value="updateModelValue({ key: field.name, value: $event })" />
-            </slot>
-          </div>
+          </slot>
         </div>
       </component>
     </div>
@@ -81,8 +83,7 @@ const props = defineProps({
   },
 
   useBox: {
-    type: Boolean,
-    default: true
+    type: Boolean
   }
 })