+++ ui/src/components/grid-generator/QasGridGenerator.vue
@@ -2,23 +2,43 @@
   <div :class="classes">
     <div v-for="(field, key) in fieldsByResult" :key="key" :class="getContainerClass({ key })">
       <slot :field="field" :name="`field-${field.name}`">
-        <slot :field="field" name="header">
-          <header :class="headerClass" :data-cy="`grid-generator-${field.name}-field`" :title="getTitle(field, 'label')">
-            {{ field.label }}
-          </header>
-        </slot>
+        <qas-item>
+          <template #header>
+            <slot :field="field" :name="`header-field-${field.name}`">
+              <slot :field="field" name="header">
+                <div :class="headerClass" :data-cy="`grid-generator-${field.name}-field`" :title="getTitle(field, 'label')">
+                  {{ field.label }}
+                </div>
+              </slot>
+            </slot>
+          </template>
 
-        <slot :field="field" name="content">
-          <div :class="contentClass" :data-cy="`grid-generator-${field.name}-result`" :title="getTitle(field, 'formattedResult')">
-            {{ field.formattedResult }}
-          </div>
-        </slot>
+          <template #content>
+            <slot :field="field" :name="`content-field-${field.name}`">
+              <slot :field="field" name="content">
+                <div :class="contentClass" :data-cy="`grid-generator-${field.name}-result`" :title="getTitle(field, 'formattedResult')">
+                  {{ field.formattedResult }}
+                </div>
+              </slot>
+            </slot>
+          </template>
+        </qas-item>
+
+        <!-- <slot :field="field" :name="`content-field-${field.name}`">
+          <slot :field="field" name="content">
+            <div class="text-body1 text-grey-10" :class="contentClass" :data-cy="`grid-generator-${field.name}-result`" :title="getTitle(field, 'formattedResult')">
+              {{ field.formattedResult }}
+            </div>
+          </slot>
+        </slot> -->
       </slot>
     </div>
   </div>
 </template>
 
 <script setup>
+import QasItem from './QasItem.vue'
+
 import useGenerator, { baseProps } from '../../composables/private/use-generator'
 import { isEmpty, humanize } from '../../helpers'
 import { useScreen } from '../../composables'