+++ ui/src/components/grid-generator/QasItem.vue
@@ -1,0 +1,31 @@
+<template>
+  <div>
+    <header class="text-caption">
+      <slot name="header">
+        {{ props.label }}
+      </slot>
+    </header>
+
+    <div class="text-body1 text-grey-10">
+      <slot name="content">
+        {{ props.value }}
+      </slot>
+    </div>
+  </div>
+</template>
+
+<script setup>
+defineOptions({ name: 'QasItem' })
+
+const props = defineProps({
+  label: {
+    type: String,
+    default: ''
+  },
+
+  value: {
+    type: [String, Number, Boolean],
+    default: ''
+  }
+})
+</script>